% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_taxonomy.R
\name{build_taxonomy}
\alias{build_taxonomy}
\title{Parse the taxon information from a Camera Trap Data Package object}
\usage{
build_taxonomy(x)
}
\arguments{
\item{x}{}
}
\value{
A tibble with the taxonomic information from a Camera Trap Data
Package
}
\description{
Parse the taxon information from a Camera Trap Data Package object
}
