% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_camtrapdp.R
\name{check_camtrapdp}
\alias{check_camtrapdp}
\title{Check a Camera Trap Data Package object}
\usage{
check_camtrapdp(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
\code{x} invisibly or error.
}
\description{
Checks if an object is a Camera Trap Data Package object with the required
properties.
}
\examples{
x <- example_dataset()
check_camtrapdp(x) # Invisible return of x if valid
}
\concept{check functions}
