% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{contributors}
\alias{contributors}
\title{Get contributors as a data frame}
\usage{
contributors(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the contributors.
}
\description{
Gets contributors from the \code{x$contributors} property in a Camera Trap Data
Package object and returns it as a data frame.
}
\examples{
x <- example_dataset()
contributors(x)
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{individuals}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
