% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert a dataset to the latest version of Camtrap DP}
\usage{
convert(dataset)
}
\arguments{
\item{dataset}{Camera Trap Data Package, as returned by \code{read_camtrap_dp()}.}
}
\value{
Converted dataset.
}
\description{
Converts a dataset that uses an older version of Camtrap DP to the latest
version (currently \code{1.0}).
}
