% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployments}
\alias{deployments}
\alias{deployments<-}
\title{Get deployments}
\usage{
deployments(x)

deployments(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{value}{data.frame to assign to the deployments of a Camera Trap Data Package Object.}
}
\value{
\code{\link[=tibble]{tibble()}} data frame with deployments.
}
\description{
Gets the deployments from a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
# Get the deployments of x
deployments(x)
# Change the deployments of x
deployments(x) <- tail(deployments(x),-1)
}
\seealso{
Other accessor functions: 
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()},
\code{\link{version}()}
}
\concept{accessor functions}
