% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Read the example data}
\usage{
example_dataset()
}
\value{
Camera Trap Data Package object.
}
\description{
This function is a helper that saves you from having to remember the url to
the example dataset based on the Mica data.
}
\examples{
example_dataset()
}
