% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_deployments.R
\name{filter_deployments}
\alias{filter_deployments}
\title{Filter deployments}
\usage{
filter_deployments(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{...}{Filtering conditions, see \code{dplyr::filter()}.}
}
\value{
Camera Trap Data Package object.
}
\description{
Subsets deployments in a Camera Trap Data Package object, retaining all
rows that satisfy a condition.
\itemize{
\item Deployments are filtered on the conditions.
\item Media are filtered on associated \code{deploymentID}.
\item Observations are filtered on associated \code{deploymentID}.
}
}
\examples{
x <- example_dataset()

# Filtering returns x (the dataset)
filter_deployments(x, deploymentID == "62c200a9")

# Use deployments() to get the filtered deployments
x \%>\%
  filter_deployments(deploymentID == "62c200a9") \%>\%
  deployments()

# Filtering by multiple criteria
x \%>\%
  filter_deployments(latitude > 51.0 & longitude > 5.0) \%>\%
  deployments()

# Associated media and observations are filtered as well
x_filtered <- filter_deployments(x, deploymentID == "62c200a9")
media(x_filtered)
observations(x_filtered)
}
\seealso{
Other filter functions: 
\code{\link{filter_media}()},
\code{\link{filter_observations}()}
}
\concept{filter functions}
