% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_media.R
\name{filter_media}
\alias{filter_media}
\title{Filter based on media}
\usage{
filter_media(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{...}{filtering expressions. Same behavior as dplyr's \code{filter()}.}
}
\value{
Camera Trap Data Package object.
}
\description{
Filter a Camera Trap Data Package object based on media. Same syntax and
behavior as in dplyr's \code{filter()} function. This function does not filter
deployments, so absences are retained. It filters observations (on
\code{mediaID}), but note that not all observations have these. It filters also
observations taken by the same deployments and same time range as the
filtered media.
}
\examples{
library(dplyr)
dataset <- example_dataset()
filter_media(dataset, favorite == TRUE)

# piping is allowed
dataset \%>\%
  filter_media(
    timestamp > as.POSIXct("2020-08-04 08:01:00")
)
}
\seealso{
Other filter functions: 
\code{\link{filter_deployments}()},
\code{\link{filter_observations}()}
}
\concept{filter functions}
