% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_observations.R
\name{filter_observations}
\alias{filter_observations}
\title{Filter based on observations}
\usage{
filter_observations(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}

\item{...}{filtering expressions. Same behavior as dplyr's \code{filter()}.}
}
\value{
Camera Trap Data Package object.
}
\description{
Filter a Camera Trap Data Package object based on observations. Same syntax
and behavior as in dplyr's \code{filter()} function. This function does not filter
deployments, so absences are retained. It does not filter media either
since a media file can be used by more than one observation and observations
are not always hard linked to media.
}
\examples{
library(dplyr)
dataset <- example_dataset()
filter_observations(dataset, observationType == "animal")
filter_observations(
  dataset,
  scientificName \%in\% c("Anas platyrhynchos", "Ardea cinerea")
)
# piping is allowed
dataset \%>\%
  filter_observations(
  observationLevel == "event", observationType == "animal"
)
}
\seealso{
Other filter functions: 
\code{\link{filter_deployments}()},
\code{\link{filter_media}()}
}
\concept{filter functions}
