% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{Get Camtrap DP version}
\usage{
get_version(dataset)
}
\arguments{
\item{dataset}{Camera Trap Data Package, as returned by \code{read_camtrap_dp()}.}
}
\value{
Camtrap DP version number or \code{dataset$profile}.
}
\description{
Gets the Camtrap DP version number used by a dataset.
The version number (\code{1}, \code{1.0}, \verb{1.2.3}, etc.) is extracted from
\code{dataset$profile}, which should contain the URL of the used Camtrap DP
standard.
When no match is found, \code{dataset$profile} is returned in its entirety.
}
