% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{individuals}
\alias{individuals}
\title{Get individuals}
\usage{
individuals(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the individuals that have an
\code{individualID}, containing the following columns:
\itemize{
\item \code{individualID}
\item \code{scientificName}
\item \code{lifeStage}
\item \code{sex}
}
}
\description{
Gets the (unique) individuals from the observations of a Camera Trap Data
Package object.
}
\examples{
x <- example_dataset()
individuals(x)
}
\seealso{
Other accessor functions:
\code{\link[=contributors]{contributors()}},
\code{\link[=deployments]{deployments()}},
\code{\link[=events]{events()}},
\code{\link[=locations]{locations()}},
\code{\link[=media]{media()}},
\code{\link[=observations]{observations()}},
\code{\link[=taxa]{taxa()}}
}
\concept{accessor functions}
