% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Get locations}
\usage{
locations(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the locations, containing the
following columns:
\itemize{
\item \code{locationID}
\item \code{locationName}
\item \code{latitude}
\item \code{longitude}
\item \code{coordinateUncertainty}
}
}
\description{
Gets the (unique) locations from the deployments of a Camera Trap Data
Package object.
}
\examples{
x <- example_dataset()
locations(x)
}
\seealso{
Other accessor functions:
\code{\link[=contributors]{contributors()}},
\code{\link[=deployments]{deployments()}},
\code{\link[=events]{events()}},
\code{\link[=individuals]{individuals()}},
\code{\link[=media]{media()}},
\code{\link[=observations]{observations()}},
\code{\link[=taxa]{taxa()}}
}
\concept{accessor functions}
