% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\alias{media<-}
\title{Get media}
\usage{
media(x)

media(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{value}{data.frame to assign to the media of a Camera Trap Data Package Object.}
}
\value{
\code{\link[=tibble]{tibble()}} data frame with media.
}
\description{
Gets the media from a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
media(x)
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{observations}()},
\code{\link{taxa}()},
\code{\link{version}()}
}
\concept{accessor functions}
