% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Get observations}
\usage{
observations(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
Observations tibble.
}
\description{
Gets the observations from a Camera Trap Data Package object.
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{media}()},
\code{\link{version}()}
}
\concept{accessor functions}
