% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.camtrapdp}
\alias{print.camtrapdp}
\title{Print a Camera Trap Data Package}
\usage{
\method{print}{camtrapdp}(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{read_camtrapdp()}.}

\item{...}{Further arguments, they are ignored by this function.}
}
\value{
\code{\link[=print]{print()}} with a summary of the Camera Trap Data Package object.
}
\description{
Prints a human-readable summary of a Camera Trap Data Package, as an
extension of \code{\link[frictionless:print.datapackage]{frictionless::print.datapackage()}}.
}
\examples{
x <- example_dataset()

Print a summary
print(x)

Print a summary after filtering
filter_deployments(x, deploymentID == "62c200a9")
}
\concept{print functions}
