% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_coordinates.R
\name{round_coordinates}
\alias{round_coordinates}
\title{Round coordinates to generalize camera trap locations}
\usage{
round_coordinates(x, digits = 3)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{digits}{Number of decimal places to round coordinates to (\code{1},
\code{2} or \code{3}).}
}
\value{
\code{x} with rounded coordinates as well as updated
\code{coordinateUncertainty} (in deployments) and \code{coordinatePrecision} (in
metadata).
}
\description{
Rounds deployment coordinates to a certain number of digits to
fuzzy/generalize camera trap locations.
This function can be used before publishing data in order to protect
sensitive species and/or prevent theft of active cameras.
}
\section{Details}{

Rounding coordinates is a recommended method to generalize sensitive
biodiversity information (see
\href{https://doi.org/10.15468/doc-5jp4-5g10#s-generalization}{Section 4.2}
in Chapman 2020).
Choose a number of digits that aligns with the sensitivity of the data and
notice the effect on precision and uncertainty.
Publish the coordinates as is (i.e. do not use this function) if the data
are not sensitive.\tabular{lll}{
   sensitivity \tab digits \tab coordinatePrecision \cr
   high \tab 1 \tab 0.1 \cr
   medium \tab 2 \tab 0.01 \cr
   low \tab 3 \tab 0.001 \cr
}


For records with \code{coordinateUncertainty = NA} the function will assume the
coordinates were obtained by GPS and use \verb{30 m} as original uncertainty,
before adding uncertainty caused by rounding.
The added rounding uncertainty is dependent on latitude and follows the
categories in
\href{https://doi.org/10.15468/doc-gg7h-s853#table-uncertainty}{Table 3} in
Chapman & Wieczorek 2020.\tabular{llll}{
   latitude \tab 0.1 degree \tab 0.01 degree \tab 0.001 degree \cr
   0째 \tab 15691 m \tab 1570 m \tab 157 m \cr
   30째 \tab 14697 m \tab 1470 m \tab 147 m \cr
   60째 \tab 12461 m \tab 1246 m \tab 125 m \cr
   85째 \tab 11211 m \tab 1121 m \tab 112 m \cr
}
}

\examples{
# Round coordinates of example package to 3 digits
x <- example_dataset()
x_rounded2 <- round_coordinates(x, 2)

# coordinatePrecision is set in metadata
x_rounded2$coordinatePrecision

# coordinateUncertainty is set in data: original uncertainty (or 30) + 1470 m
x_rounded2$data$deployments$coordinateUncertainty
}
\seealso{
Other transformation functions: 
\code{\link{write_dwc}()}
}
\concept{transformation functions}
