% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa}
\alias{taxa}
\title{Get taxa}
\usage{
taxa(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the taxonomic information,
containing at least a \code{scientificName} column.
}
\description{
Gets the (unique) scientific names and associated taxonomic information from
the observations of a Camera Trap Data Package object.
Duplicate taxa (i.e. with the same \code{scientificName}) are removed, retaining
the taxon with (first) a \code{taxonID} and (second) the most taxonomic
information.
}
\examples{
x <- example_dataset()
taxa(x)
}
\seealso{
Other accessor functions:
\code{\link[=contributors]{contributors()}},
\code{\link[=deployments]{deployments()}},
\code{\link[=events]{events()}},
\code{\link[=individuals]{individuals()}},
\code{\link[=locations]{locations()}},
\code{\link[=media]{media()}},
\code{\link[=observations]{observations()}}
}
\concept{accessor functions}
