% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Get taxonomy}
\usage{
taxonomy(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
\code{tibble()} data.frame.
}
\description{
Gets the taxonomy information from the observations of a Camera Trap Data
Package object. Notice that this information could differ from the taxonomy
described in \code{taxonomic} slot.
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{version}()}
}
\concept{accessor functions}
