% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxon.R
\name{update_taxon}
\alias{update_taxon}
\title{Update taxon information}
\usage{
update_taxon(x, from, to)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}

\item{from}{\code{scientificName} to be updated.}

\item{to}{List with new taxon information. Must have a \code{scientificName}
property.}
}
\value{
\code{x} with updated taxon information.
}
\description{
Update taxon information
}
\examples{
x <- example_dataset()
from <- "Anas platyrhynchos"
to <- list(
scientificName = "Anas",
taxonID = "https://www.checklistbank.org/dataset/9910/taxon/V8R",
taxonRank = "genus",
vernacularNames.eng = "dabbling ducks"
)
}
\seealso{
Other transformation functions: 
\code{\link{merge_camtrapdp}()},
\code{\link{round_coordinates}()},
\code{\link{shift_time}()},
\code{\link{write_dwc}()},
\code{\link{write_eml}()}
}
\concept{transformation functions}
