% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_camtrapdp.R
\name{write_camtrapdp}
\alias{write_camtrapdp}
\title{Write a Camera Trap Data Package to disk}
\usage{
write_camtrapdp(x, directory)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}

\item{directory}{Path to local directory to write files to.}
}
\description{
Writes a Camera Trap Data Package to disk as a datapackage.json and CSV
files.
}
\examples{
x <- example_dataset()
x <- filter_deployments(x, deploymentID == "00a2c20d")
(write_camtrapdp(x, directory = "my_directory"))

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
