% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Camtrap DP data to Darwin Core}
\usage{
write_dwc(x, directory = ".")
}
\arguments{
\item{x}{A Camtrap DP, as read by \href{https://camtrap-dp.tdwg.org}{read_camtrapdp()}.}

\item{directory}{Path to local directory to write file(s) to.
If \code{NULL}, then a list of data frames is returned instead, which can be
useful for extending/adapting the Darwin Core mapping before writing with
\code{\link[readr:write_delim]{readr::write_csv()}}.}
}
\value{
CSV and \code{meta.xml} files written to disk or a list of data
frames when \code{directory = NULL}.
}
\description{
Transforms data from a \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package} to \href{https://dwc.tdwg.org/}{Darwin Core}.
The resulting CSV files can be uploaded to an \href{https://www.gbif.org/ipt}{IPT} for publication to GBIF.
A \code{meta.xml} file is included as well.
}
\section{Transformation details}{

Data are transformed into an
\href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence core} and
\href{https://ac.tdwg.org/termlist/}{Audiovisual core}.
This \strong{follows recommendations} discussed and created by Peter Desmet,
John Wieczorek, Lien Reyserhove, Ben Norton and others.

The following terms are set from the \code{x} metadata:
\itemize{
\item \strong{datasetName}: Title as provided in \code{x$title}.
\item \strong{datasetID}: Identifier as provided in \code{x$id}.
Can be a DOI.
\item \strong{rightsHolder}: Rights holder with role \code{rightsHolder} as provided in
\code{x$contributors}. If no rightsHolder is assigned, rightsHolder is set to the
organization of the first contributor.
\item \strong{collectionCode}: Platform name as provided in \code{x$sources[[1]]$title}.
\item \strong{license}: License with scope \code{data} as provided in \code{x$licenses}.
\item \strong{rights} for media files: License with scope \code{media} as provided in
\code{x$licenses}.
\item \strong{dwc:dataGeneralizations}: "coordinates rounded to
\code{x$coordinatePrecision} degree".
\item \strong{coordinatePrecision}: \code{x$coordinatePrecision} (e.g. \code{0.001}).
}

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (of camera traps) are parent events, with observations
(machine observations) as child events. No information about the parent
event is provided other than its ID, meaning that data can be expressed in
an Occurrence Core with one row per observation and \code{parentEventID} shared
by all occurrences in a deployment.
\item Sequence-based observations share an \code{eventID} per sequence, image-based
observations share an \code{eventID} per image.
\item The image(s) an observation is based on are provided in the
\href{https://ac.tdwg.org/termlist/}{Audiovisual core}, with a foreign
key to the observation.
\item Excluded are records that document blank or unclassified media, vehicles
and observations of humans.
}
}

\examples{
\dontrun{
x <- example_dataset()
write_dwc(x)
}
}
\concept{publication functions}
