% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{Transform Camtrap DP metadata to EML}
\usage{
write_eml(x, directory)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{directory}{Path to local directory to write file to.}
}
\value{
\code{eml.xml} file written to disk. And invisibly, an \code{EML} object is
returned.
}
\description{
Transforms the metadata of a \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package} to an \href{https://eml.ecoinformatics.org/}{EML}
file that can be uploaded to a \href{https://www.gbif.org/ipt}{GBIF IPT} for
publication.
}
\section{Transformation details}{

Metadata is derived from what is provided in \code{x}.
The following properties are set:
\itemize{
\item \strong{title}: Title as provided in \code{x$title}.
\item \strong{description}: Description as provided in \code{x$description}.
The description is followed by paragraph how the data have been
standardized.
\item \strong{license}: License with scope \code{data} as provided in \code{x$licenses}.
\item \strong{creators}: Contributors (all roles) as provided in \code{x$contributors}.
\item \strong{contact}: First creator.
\item \strong{metadata provider}: First creator.
\item \strong{keywords}: Keywords as provided in \code{keywords}.
\item \strong{geographic coverage}: Bounding box as provided in \code{x$spatial}.
\item \strong{taxonomic coverage}: Species (no other ranks) as provided in
\code{x$taxonomic}.
\item \strong{temporal coverage}: Date range as provided in \code{x$temporal}.
\item \strong{project data}: Title, acronym as identifier, description, and sampling
design as provided in \code{x$project}.
\item \strong{alternative identifier}: Identifier as provided in \code{x$id}.
If this is a DOI, no new DOI will be created when publishing to GBIF.
\item \strong{external link}: URL of the project as provided in
\code{x$project$path}.
}

To be set manually in the GBIF IPT: \strong{type}, \strong{subtype},
\strong{update frequency} and \strong{publishing organization}.

Not set: \strong{sampling methods} and \strong{citations}.

Not applicable: \strong{collection data}.
}

\examples{
\dontrun{
x <- example_dataset()
write_eml(x, directory = "my_directory")

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
}
\seealso{
Other transformation functions: 
\code{\link{write_dwc}()}
}
\concept{transformation functions}
