% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert a Camera Trap Data Package}
\usage{
convert(package)
}
\arguments{
\item{package}{Camera Trap Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
A \code{camtrapdp} object.
}
\description{
Converts a Camera Trap Data Package that uses an older version of Camtrap DP
to the latest version (currently \code{1.0}).
}
