% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_time.R
\name{correct_time}
\alias{correct_time}
\title{Corrects date-time}
\usage{
correct_time(x, deploymentID, duration)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{read_camtrapdp()}.}

\item{deploymentID}{One or more deploymentID's.}

\item{duration}{Date-time difference between the wrong and right time.}
}
\value{
\code{x} with date-times corrected.
}
\description{
Corrects date-times in a Camera Trap Data package object. This function can
be used when the time settings of one or more deployments were wrong and need
to be corrected afterwards.
}
\examples{
x <- example_dataset()
deploymentID <- c("00a2c20d", "29b7d356")
# calculate duration
library(lubridate)
wrong <- ymd_hms("2024-04-01T00:00:00", tz = "UTC")
right <- ymd_hms("2024-04-01T02:00:00", tz = "UTC")
duration <- as.duration(interval(wrong, right))
# correct time
x_corrected <- correct_time(x, deploymentID, duration)
# inspect results
deployments(x_corrected)
}
\seealso{
Other transformation functions: 
\code{\link{round_coordinates}()},
\code{\link{write_dwc}()}
}
\concept{transformation functions}
