% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployments}
\alias{deployments}
\title{Get deployments}
\usage{
deployments(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
\code{\link[=tibble]{tibble()}} data frame with deployments.
}
\description{
Gets the deployments from a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
deployments(x)
}
\seealso{
Other accessor functions: 
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()},
\code{\link{version}()}
}
\concept{accessor functions}
