% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Read the Camtrap DP example dataset}
\usage{
example_dataset()
}
\value{
Camera Trap Data Package object.
}
\description{
Reads the \href{https://camtrap-dp.tdwg.org/example/}{Camtrap DP example dataset}.
This dataset is maintained and versioned with the Camtrap DP standard.
}
\concept{sample data}
