% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_deployments}
\alias{filter_deployments}
\title{Define filter method for camtrapdp class}
\usage{
filter_deployments(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
Camera Trap Data Package object.
}
\description{
Internal function for filtering deployments
}
\examples{
# example code
dataset <- example_dataset()
filter_deployments(dataset, locationName == "B_HS_val 2_processiepark")
}
\concept{filter functions}
