% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_deployments.R
\name{filter_deployments}
\alias{filter_deployments}
\title{Filter based on deployments}
\usage{
filter_deployments(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{...}{filtering expressions. Same behavior as dplyr's \code{filter()}.}
}
\value{
Camera Trap Data Package object.
}
\description{
Filter a Camera Trap Data Package object based on deployments. Same syntax
and behavior as in dplyr's \code{filter()} function. The deployments will be filtered based on the provided conditions.
The media and the observations of the removed deployments will be removed as well.
}
\examples{
library(dplyr)
dataset <- example_dataset()
filter_deployments(dataset, locationName == "B_HS_val 2_processiepark")
filter_deployments(dataset, deploymentStart > as.POSIXct("2021-01-01"))
# piping is allowed
dataset \%>\% filter_deployments(latitude > 51, longitude < 5)
}
\seealso{
Other filter functions: 
\code{\link{filter_media}()},
\code{\link{filter_observations}()}
}
\concept{filter functions}
