% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{Get the Camtrap DP version used by a dataset}
\usage{
get_version(dataset)
}
\arguments{
\item{dataset}{Camera Trap Data Package, as returned by \code{read_camtrap_dp()}.}
}
\value{
Camtrap DP version number (e.g. \code{1.0}) or \code{dataset$profile}.
}
\description{
Extracts the Camtrap DP version number used by a dataset, by searching for
\verb{camtrap-dp/<version-number>} in \code{dataset$profile}, which is expected to
contain the URL of the used Camtrap DP standard.
When no match is found, \code{dataset$profile} is returned in its entirety.
}
