% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Get locations}
\usage{
locations(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
\code{tibble()} data frame with the following columns:
\itemize{
\item \code{locationID}
\item \code{locationName}
\item \code{latitude}
\item \code{longitude}
\item \code{coordinateUncertainty}
}
}
\description{
Extract unique locations of the deployments of a Camera Trap Data Package
object.
}
\examples{
dataset <- example_dataset()
locations(dataset)
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{version}()}
}
\concept{accessor functions}
