% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Get locations}
\usage{
locations(x)
}
\arguments{
\item{x}{a Camera Trap Data Package object.}
}
\value{
a tibble data.frame with the following columns:
\itemize{
\item \code{locationID}
\item \code{locationName}
\item \code{latitude}
\item \code{longitude}
\item \code{coordinateUncertainty}
}
}
\description{
Extract unique locations of the deployments of a Camera Trap Data Package
object.
}
\examples{
dataset <- example_dataset()
locations(dataset)
}
