% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\alias{media<-}
\title{Get media}
\usage{
media(x)

media(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{value}{data.frame to assign to the media of a Camera Trap Data Package Object.}
}
\value{
\code{\link[=tibble]{tibble()}} data frame with media.
}
\description{
Gets the media from a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
# Get the media of x
media(x)
# Change the media of x
media(x) <-
  data.frame(
    stringsAsFactors = FALSE,
    mediaID = c("5b0ccf8a", "1640ad29"),
    deploymentID = c("577b543a", "00a2c20d"),
    timestamp = c("2020-06-23 23:33:56", "2020-05-31 20:07:31"),
    filePath = c("https://multimedia.agouti.eu/assets/5b0ccf8a-4696-4814-8bb6-c96bb248bc3f/file",
                 "https://multimedia.agouti.eu/assets/1640ad29-e2cb-43cc-8cd6-c1ef2ffaadd7/file"),
    filePublic = c(TRUE, TRUE),
    fileName = c("20200811211208-RCNX0061.JPG",
                 "20200709093338-RCNX0043.JPG"),
    fileMediatype = c("image/jpeg", "image/jpeg"),
    exifData = c(NA, NA),
    favorite = c(NA, NA),
    mediaComments = c(NA, NA),
    eventID = c("b4b39b00", "45abeadc"),
    captureMethod = as.factor(c("activityDetection", "activityDetection"))
   )

}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{observations}()},
\code{\link{taxa}()},
\code{\link{version}()}
}
\concept{accessor functions}
