% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{Get media}
\usage{
media(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
\code{tibble()} data frame with media.
}
\description{
Gets the media from a Camera Trap Data Package object.
}
\examples{
dataset <- example_dataset()
media(dataset)
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{observations}()},
\code{\link{version}()}
}
\concept{accessor functions}
