% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{Get media}
\usage{
media(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.
Also works on a Frictionless Data Package, as returned by
\code{frictionless::read_package()}.}
}
\value{
Media tibble.
}
\description{
Gets the media from a Camera Trap Data Package object.
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{observations}()},
\code{\link{taxonomy}()},
\code{\link{version}()}
}
\concept{accessor functions}
