% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_camtrapdp.R
\name{merge_camtrapdp}
\alias{merge_camtrapdp}
\title{Merge camtrapdp packages}
\usage{
merge_camtrapdp(x1, x2, name, title)
}
\value{
\code{x}
}
\description{
Merge camtrapdp packages
}
\examples{
x1 <- example_dataset() \%>\%
  filter_deployments(deploymentID \%in\% c("00a2c20d", "29b7d356"))
x2 <- example_dataset() \%>\%
  filter_deployments(deploymentID \%in\% c("577b543a", "62c200a9"))
x_merged <- merge_camtrapdp(x1, x2, "new package name", "new title")
}
