% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\alias{observations<-}
\title{Get observations}
\usage{
observations(x)

observations(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{value}{data.frame to assign to the observations of a Camera Trap Data Package Object.}
}
\value{
\code{\link[=tibble]{tibble()}} data frame with observations.
}
\description{
Gets the observations from a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
# Get the observations of x
observations(x)
# Change the observations of x
observations(x) <-
  observations(x)[observations(x)$observationLevel == "media", ]
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{taxa}()},
\code{\link{version}()}
}
\concept{accessor functions}
