% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.camtrapdp}
\alias{print.camtrapdp}
\title{Print a Camera Trap Data Package}
\usage{
\method{print}{camtrapdp}(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by
\code{read_camtrapdp()}.}

\item{...}{}
}
\value{
\code{\link[=print]{print()}} with a summary of the Camera Trap Data Package object.
}
\description{
Prints a human-readable summary of a Camera Trap Data Package, as an
extension of the printing provided by the frictionless package. This summary
provides the most important aspects of a Camera Trap Data Package.
}
\examples{
dp <- example_dataset()
print(example_dataset)
}
