% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrapdp.R
\name{read_camtrapdp}
\alias{read_camtrapdp}
\title{Read a Camera Trap Data Package}
\usage{
read_camtrapdp(file, media_eventid = TRUE)
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}

\item{media_eventid}{Whether \code{eventID}s found in observations should be
assigned to media.}
}
\value{
Camera Trap Data Package object.
}
\description{
Reads files from a \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package (Camtrap DP)} into memory.
}
\concept{read functions}
