% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa}
\alias{taxa}
\title{Get taxa}
\usage{
taxa(x)
}
\value{
\code{\link[=tibble]{tibble()}} data frame with the taxonomic information, containing at
least a \code{scientificName} column.
}
\description{
Gets the (unique) scientific names and associated taxonomic information from
the observations of a Camera Trap Data Package object.
}
\examples{
x <- example_dataset()
taxa(x)
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()}
}
\concept{accessor functions}
