% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxon.R
\name{update_taxon}
\alias{update_taxon}
\title{Update taxon information}
\usage{
update_taxon(x, from, to)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}

\item{from}{\code{scientificName} to be updated.}

\item{to}{List with new taxon information. It must have a \code{scientificName}
property.}
}
\value{
\code{x} with updated taxon information.
}
\description{
Updates taxon information in both data and metadata.

If \code{to}is a list containing a scientificName that already exists in the
dataset, duplicate taxa (i.e. with the same \code{scientificName}) are removed.
The function retains the taxon with (first) a \code{taxonID} and (second) the most
taxonomic information.

Taxonomic information for a \code{scientificName} can be updated by replacing the
existing taxon with the same \code{scientificName} but different taxonomic details
provided in \code{to}.
}
\examples{
# Example 1: change scientificName
x <- example_dataset()
from <- "Ardea cinerea"
to <- list(
scientificName = "Ardea",
taxonID = "https://www.checklistbank.org/dataset/COL2023/taxon/32FH"
)
x_updated <- update_taxon(x, from, to)
taxa(x_updated)

# Example 2: update taxon information
from <- "Anas platyrhynchos"
to <- list(
scientificName = "Anas platyrhynchos",
taxonID = "https://www.checklistbank.org/dataset/COL2023/taxon/DGP6",
taxonRank = "species",
vernacularNames.fr = "canard chipeau"
)

x_updated <- update_taxon(x, from, to)
taxa(x_updated)
}
\seealso{
Other transformation functions: 
\code{\link{merge_camtrapdp}()},
\code{\link{round_coordinates}()},
\code{\link{shift_time}()},
\code{\link{write_dwc}()},
\code{\link{write_eml}()}
}
\concept{transformation functions}
