% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxon.R
\name{update_taxon}
\alias{update_taxon}
\title{Update taxon}
\usage{
update_taxon(x, from, to)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[=read_camtrapdp]{read_camtrapdp()}}.}

\item{from}{\code{scientificName} of the taxon to update.}

\item{to}{Named list with taxon information, e.g.
\code{list(scientificName = "Ardea", taxonRank = "genus", vernacularname.eng = "great herons")}.}
}
\value{
\code{x} with updated taxon information.
}
\description{
Updates information for a taxon in data and metadata.
This allows:
\enumerate{
\item Taxon update: when \code{to} and \code{from$scientificName} are the same, then the
existing taxon is updated with the provided information.
\item Taxon rename: when \code{to} and \code{from$scientificName} are different and the
latter is new, then the existing taxon is renamed and updated with the
provided information.
\item Taxon lump: when \code{to} and \code{from$scientificName} are different and the
latter is already present, then the existing taxon is lumped into the already
present taxon.
Information is retained from the taxon (provided or already present) with
(first) a \code{taxonID} and (second) the most information.
}
}
\examples{
x <- example_dataset()

# Update taxonomic information for "Anas platyrhynchos"
updated_x <- update_taxon(
  x,
  from = "Anas platyrhynchos",
  to = list (
    scientificName = "Anas platyrhynchos",
    taxonID = "https://www.checklistbank.org/dataset/COL2023/taxon/DGP6",
    taxonRank = "species",
    vernacularNames.fra = "canard colvert"
  )
)

# Lump "Ardea cinerea" into already present taxon "Ardea".
# Provided information is ignored, because the present taxon has a taxonID.
updated_x <- update_taxon(
  x,
  from = "Ardea cinerea",
  to = list(scientificName = "Ardea", vernacularname.fra = "grands hérons")
)
}
\seealso{
Other transformation functions: 
\code{\link{merge_camtrapdp}()},
\code{\link{round_coordinates}()},
\code{\link{shift_time}()},
\code{\link{write_dwc}()},
\code{\link{write_eml}()}
}
\concept{transformation functions}
